requirements.txt
fastapi==0.88.0
psycopg2-binary==2.9.5
SQLAlchemy==1.4.45
uvicorn==0.20.0
python-dotenv==0.10.5
